// izadji/src/utils/enums/pageLinkComponents.ts
enum PageLinkComponents {
  ABOUT_US = 'page-link-components.about-us',
  BLOG = 'page-link-components.blog',
  BLOG_NEWS_PAGE = 'page-link-components.blog-news-page',
  CONTACT_PAGE = 'page-link-components.contact-page',
  DOCUMENTATION_PAGE = 'page-link-components.documentation-page',
  DONATE_PAGE = 'page-link-components.donate-page',
  HOMEPAGE = 'page-link-components.homepage',
  PRIDE_PAGE = 'page-link-components.pride-page',
  PUBLICATION_PAGE = 'page-link-components.publication-page',
}

export default PageLinkComponents;